Migrate to shared GitHub Actions workflow
Migrates to a new, shared GitHub Actions workflow template provided by the BerkeleyLibrary/.github repository. This does essentially what we have always done, but pulls the workflow configuration from the central repository. The CI process is as follows:

- Build: The Docker image is built, tagged (branch, SHA, and tag), and pushed to ghcr.io/berkeleylibrary/geodata.
- Test: The built image is tested using `docker compose`.

The test stage depends on the consuming repository having implemented `bin/test`, `bin/setup`, and `docker-compose(.ci)?.yml` files. For more information on that, see the accompanying files in this commit or the forthcoming documentation in the shared workflows repo.