package: Replace rollup for @rollup/wasm-node
The wasm-node version of Rollup is portable to any Node platform, so it
does not have per-CPU and per-OS native builds.  By using this as our
implementation of Rollup, we avoid having to add each platform/OS we
want to run Geodata on.  (We can also now run on places Rollup doesn't.)

Closes: AP-344